# Faster array parsing